Memcached request steering
This patch adds an XDP program to steer Memcached requests to the right
core.

The keyspace is partitioned between CPU cores so that each core is
responsible for a subset of the keys. Keys are partitioned using a
Murmur3 hash of the key modulo number of cores. Partitioning of keyspace
also requires requests to be forwarded to the right core. However, the
NIC distributes packets to RX queues by flow, which means our request is
likely to arrive on a wrong core.

We could forward the request in userspace, but that has a high overhead
because (1) packet has to run through the Linux kernel network stack and
(2) inter-thread and inter-process communication is expensive in
userspace because it requires a wakeup. Therefore, let's parse the
Memcache (binary protocol) request in an XDP program, determine the
correct partition of the key, and use XDP_REDIRECT with CPU bitmap to
steer the request to the correct core.

Note: I have verified that this builds with the XDP toolchain, but I
have not loaded or actually executed any of this.